Update.
	* libio/tst-freopen.c (main): Use correct formats for __LINE__.
	* libio/tst-mmap-setvbuf.c (main): Likewise.
	* libio/tst-widetext.c (main): Likewise.

	* libio/tst-mmap-eofsync.c (do_test): Use correct format in error
	message.
	* libio/tst-mmap-fflushsync.c (do_test): Likewise.
	* libio/tst-mmap2-eofsync.c (do_test): Likewise.

+++ linuxthreads/tst-context.c
@@ -98,7 +98,7 @@ main (void)
     }
 
   for (n = 0; n < N; ++n)
-    if (pthread_create (&th[n], NULL, threadfct, (void *) n) != 0)
+    if (pthread_create (&th[n], NULL, threadfct, (void *) (long int) n) != 0)
       error (EXIT_FAILURE, errno, "cannot create all threads");
 
   for (n = 0; n < N; ++n)
