+++ localedata/tests-mbwc/tst_wcscpy.c
@@ -47,7 +47,8 @@ tst_wcscpy (FILE * fp, int debug_flg)
 		{
 		  fprintf (stderr,
 			   "ws1[ %d ] = 0x%lx <-> wx_ex[ %d ] = 0x%lx\n", i,
-			   ws1[i], i, ws_ex[i]);
+			   (unsigned long int) ws1[i], i,
+			   (unsigned long int) ws_ex[i]);
 		}
 
 	      if (ws1[i] != ws_ex[i])
