+++ localedata/tests-mbwc/tst_wcsstr.c
@@ -32,7 +32,7 @@ tst_wcsstr (FILE * fp, int debug_flg)
 	    {
 	      fprintf (stderr,
 		       "	ret[ 0 ] = 0x%lx <-> 0x%lx = ws2[ 0 ]\n",
-		       ret[0], ws2[0]);
+		       (unsigned long int) ret[0], (unsigned long int) ws2[0]);
 	    }
 	}
 
@@ -60,7 +60,8 @@ tst_wcsstr (FILE * fp, int debug_flg)
 	      {
 		fprintf (stderr,
 			 "	: ret[ %d ] = 0x%lx <-> 0x%lx = ws2[ %d ]\n",
-			 i, ret[i], ws2[i], i);
+			 i, (unsigned long int) ret[i],
+			 (unsigned long int) ws2[i], i);
 	      }
 
 	    if (ret[i] != ws2[i])
